% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AG_gene_list}
\alias{AG_gene_list}
\title{Angiogenesis signature genes}
\format{
A vector of characters with length 41:
\describe{
\item{\code{Gene}}{character Gene hugo symbol}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3743050/}
}
\usage{
AG_gene_list
}
\description{
A set of 41 unique gene Hugo IDs as the core human Angiogenesis signature
}
\keyword{datasets}
