% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_syng_score.R
\name{calculate_syng_score}
\alias{calculate_syng_score}
\title{Calculate four synergy scores given median and standard error of means of four groups}
\usage{
calculate_syng_score(
  m_LL,
  m_LH,
  m_HL,
  m_HH,
  SEM2_LL,
  SEM2_LH,
  SEM2_HL,
  SEM2_HH,
  method
)
}
\arguments{
\item{m_LL}{median in Low_Low expression group}

\item{m_LH}{median in Low_High expression group}

\item{m_HL}{median in High_Low expression group}

\item{m_HH}{median in High_High expression group}

\item{SEM2_LL}{Standard error of the mean of the representing group}

\item{SEM2_LH}{Standard error of the mean of the representing group}

\item{SEM2_HL}{Standard error of the mean of the representing group}

\item{SEM2_HH}{Standard error of the mean of the representing group}

\item{method}{A character string indicating which synergy score to be used. one of "max" or "independence".
Default is "max".}
}
\value{
A numerical  scores
}
\description{
Calculate four synergy scores given median and standard error of means of four groups
}
\details{
Calculates a synergy score given median and standard error of means of four groups each representing different expression levels of two genes.

Synergy scores are defined as the deviation of median value in high-high group from the its expected median as estimated using one of the "max" or "independence" models. Lets define ma and mb as the difference between the levels of immune feature in low-low group and low-high/high-low groups respectively. The expected median of high-high group is then defined as max(ma , mb) in "max" model and as ma+mb-ma*mb in "independence" model. Synergy score is then calculated as the difference between mc and the expected median. Scores are called synergistic if mc exceeds the expected median. Antagonistic scores are ignored. The value of the synergy score are multiplied by 100 and a sign factor depending on the direction of the effect. By default synergy score will be zero if median values change in opposite directions. Differences in median values that are smaller that combined standard errors are considered as zero.
}
\keyword{score}
\keyword{synergy}
