% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMT_gene_list}
\alias{EMT_gene_list}
\title{EMT signature genes}
\format{
A data frame with 77 rows and 3 columns:
\describe{
\item{\code{genes}}{character Gene hugo symbol}
\item{\code{group}}{character Epithelial vs Mesenchymal}
\item{\code{sign}}{numeric Negative for Epithelial and Positive for Mesenchymal}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4737991/#SD1}
}
\usage{
EMT_gene_list
}
\description{
A set of 77 unique genes as the pan-cancer EMT signature
}
\keyword{datasets}
