% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angio_score.R
\name{get_angio_score}
\alias{get_angio_score}
\title{Calculates Angiogenesis score}
\usage{
get_angio_score(pdf)
}
\arguments{
\item{pdf}{A numeric data frame or matrix with samples as columns and gene/proteins as rows.}
}
\value{
a dataframe of Angiogenesis scores for each sample
}
\description{
Calculates Angiogenesis score
}
\details{
The angiogenesis score is defined as mean value of gene expressions in AG_gene_list. Scores are mean-centered and normalized by standard deviation based on distribution of scores in all samples.
}
\examples{
get_angio_score(pdf=sample_mRNA_data)
}
\keyword{Angiogenesis,}
\keyword{pathway}
\keyword{score}
