% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CScore.R
\name{Get_CScore}
\alias{Get_CScore}
\title{calculates a coopretivity score}
\usage{
Get_CScore(fdata)
}
\arguments{
\item{fdata}{a formatted dataframe}
}
\value{
a dataframe with gene names and synergy scores
}
\description{
Generates coopretivity boxplots immune checkpoints
}
\examples{
im_syng(fdata=myformatteddata)
Get_CScore()
}
\keyword{synergy}
