% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emt_score.R
\name{get_emt_score}
\alias{get_emt_score}
\title{Calculates EMT score
Calculates EMT score from a gene list}
\usage{
get_emt_score(pdf)
}
\arguments{
\item{pdf}{a formated mRNA data frame}
}
\value{
a dataframe of correlation coefficient and p.values
}
\description{
Calculates EMT score
Calculates EMT score from a gene list
}
\examples{
get_emt_score(pdf=sample_mRNA_data)
}
\keyword{checkpoints}
\keyword{immune}
