% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_EMTscore.R
\name{Get_EMTscore}
\alias{Get_EMTscore}
\title{Calculates pathway score
Calculates a pathway score from a gene list}
\usage{
Get_EMTscore(pdf)
}
\arguments{
\item{pdf}{a formated mRNA data frame}
}
\value{
a dataframe of correlation coefficient and p.values
}
\description{
Calculates pathway score
Calculates a pathway score from a gene list
}
\examples{
Get_EMTscore(pdf=sample_mRNA_data)
Get_EMTscore()


}
\keyword{checkpoints}
\keyword{immune}
