% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Calculates and transforms default immune features within a specified sample cohort .}
\usage{
get_features(pdf)
}
\arguments{
\item{pdf}{A numeric data frame or matrix with samples as columns and gene/proteins as rows.}
}
\value{
A list with two dataframes containing transformed values for  immune features and immune cell type fractions as listed in \code{TCGA_immune_features_list}.
}
\description{
Calculates and transforms default immune features within a specified sample cohort .
}
\details{
Given a cohort of samples, calculates immune features within the cohort and transforms their values using probability integral transform. The resulting feature values will have a standard uniform distribution between \link{0,1}
}
\examples{
get_features(pdf=sample_mRNA_data)

}
\keyword{Immune}
\keyword{Probability}
\keyword{feature,}
\keyword{integral}
\keyword{transform}
