% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifng_score.R
\name{get_ifng_score}
\alias{get_ifng_score}
\title{Calculates Z score for IFNG
Calculates Z score, the mean-centered and normalized by standard deviation
expression values for IFNG within the specified dataset}
\usage{
get_ifng_score(pdf)
}
\arguments{
\item{pdf}{a formated mRNA data frame}
}
\value{
a dataframe of correlation coefficient and p.values
}
\description{
Calculates Z score for IFNG
Calculates Z score, the mean-centered and normalized by standard deviation
expression values for IFNG within the specified dataset
}
\examples{
get_ifng_score(pdf=sample_mRNA_data)
}
\keyword{checkpoints}
\keyword{immune}
