% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_quantile_rank.R
\name{Get_qunatile_rank}
\alias{Get_qunatile_rank}
\title{A correlation function}
\usage{
Get_qunatile_rank(df)
}
\arguments{
\item{df}{a numeric dataframe}
}
\value{
a list of multiple dataframes of correlation coefficients and p.values
}
\description{
calculates quantile ranking for each column of a datarfame
}
\examples{
im_coop( df=sample_mRNA_data)
Get_qunatile_rank()

}
\keyword{quantile}
\keyword{ranking}
