% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sif.R
\name{get_sif}
\alias{get_sif}
\title{Generates SIF formatted data-frame}
\usage{
get_sif(df, cohort, Immune_Feature, cutoff)
}
\arguments{
\item{df}{a data frame of synergy scores as outputed by im_syng}

\item{cohort}{a charachter string indicating a single cohort}

\item{Immune_Feature}{a charachter string indicating a single immune feature}

\item{cutoff}{a numeric value indicating the minimum cutoff value for absolute values of  synergy score. Default is 1}
}
\value{
a binary SIF as a data.frame with three columns:
"Cotarget", "Score", "Checkpoint"
}
\description{
Generates SIF formatted data-frame
}
\details{
get_sif generates a data-frame in SIF format. (See http://wiki.cytoscape.org/Cytoscape_User_Manual/Network_Formats/)
}
\examples{
df <- im_syng_tcga(cotarget = c("BRAF","AKT1"),cohort = c("acc","gbm"))
          get_sif(df=df,cohort="acc", Immune_Feature="B.Cells.Naive"   )
}
\keyword{SIF}
