% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_syng_score.R
\name{get_syng_score}
\alias{get_syng_score}
\title{Calculate synergy score for the association effect of two genes/proteins on a feature or response}
\usage{
get_syng_score(fdata, method)
}
\arguments{
\item{fdata}{A numeric data frame with 3 columns: A numeric value of an immune feature and two group ID's based on quantile stratification of two genes/proteins.}

\item{method}{a charachter string indicating which synergy score to be used. one of "HSA" or "Bliss". Default is HSA.}
}
\value{
A data frame containing synergy scores
}
\description{
Calculate synergy score for the association effect of two genes/proteins on a feature or response
}
\details{
fdata is a numeric data frame with 3 columns: The first column contains the numeric value of a single feature. The second and third column are any of the numbers 1 or 4, representing low or high levels of expression of two genes/proteins respectely. For details of quantile startification see get_quantile_rank.

Synergy score calculation are based on Bliss defenition of independence: First median values of the feature column are determined for four groups of low-low, low-high, high-low and high-high expressions. Lets define ma, mb, and mc as median of low-high, high-low and high-high groups minus median of low-low group. The synergy score is then defined as
floor(abs(sign(ma) + sign(mb) + sign(mc)) / 3)*(abs(mc) - max(abs(ma) , abs(mb))). The value of the score is positive for synergistic effects and negative for antagonistic effects. By default the synergy score will be zero if ma, mb , and mc have opposite signs.
}
\examples{
dft <- TCGA_EMT$EMTscore
mydata<- cbind(feature=dft,gene1=sample(c(1,4),length(dft),replace = T),gene2=sample(c(1,4),length(dft),replace = T))
get_syng_score(mydata)
}
\keyword{Bliss}
\keyword{HSA,}
\keyword{scoring,}
\keyword{synergy}
