% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icp_gene_list}
\alias{icp_gene_list}
\title{Default list of immune checkpoints}
\format{
A vector of characters
\describe{
Each character is a gene hugo symbol
}
}
\source{
\url{https://doi.org/10.1101/2021.10.06.462889}
}
\usage{
icp_gene_list
}
\description{
List of 50 genes including 29 immune checkpoints and their 21 interacting ligands/receptors
}
\keyword{datasets}
