% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_boot.R
\name{im_boot}
\alias{im_boot}
\title{A correlation function}
\usage{
im_boot(gene1, gene2, Immune_Feature, N_iteration, df_mrna, df_ict, df_lf)
}
\arguments{
\item{gene1}{a single gene name}

\item{gene2}{a single gene name}

\item{df_mrna}{a formated mRNA data frame}

\item{df_ict}{an optional formated immune cell type fractions data frame}

\item{df_lf}{an optional formated Leukocyte fraction data frame}

\item{Immune_Feautue}{an immune feature name as listed in im_syng output.}
}
\value{
a list of multiple dataframes of correlation coefficients and p.values
}
\description{
Generates coopretivity boxplots immune checkpoints
}
\examples{
im_boot(gene1 = "TGFB1",gene2="TNFSF4",
                  Immune_Feature="EMTscore",
                  N_iteration=1000,
                  df_mrna =  sample_mRNA_data,
                  df_lf = sample_Leukocyte_fraction_data,
                  df_ict = sample_immune_cell_fraction_data)
}
\keyword{boxplots}
