% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_boot_tcga.R
\name{im_boot_tcga}
\alias{im_boot_tcga}
\title{A bootstrapping function}
\usage{
im_boot_tcga(gene1, gene2, cohort, Immune_Feature, N_iteration)
}
\arguments{
\item{gene1}{a single gene name}

\item{gene2}{a single gene name}

\item{cohort}{a single TCGA disease}

\item{Immune_Feature}{an immune feature name as listed in im_syng_tcga output.}
}
\value{
the probability of observation
}
\description{
Generates a probabilistic measure for an observation useing bootstrapping
}
\examples{
im_boot_tcga(gene1 = "ACVRL1",gene2="CD274",
                  cohort="brca",
                  Immune_Feature="Leukocyte_fraction",
                  N_iteration=100000)
}
\keyword{bootstrap,}
\keyword{pvalue,}
\keyword{random}
\keyword{sampling}
