% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_boxplot.R
\name{im_boxplot}
\alias{im_boxplot}
\title{Generate a stratified boxplot for immune feature values}
\usage{
im_boxplot(onco_gene, icp_gene, data_expression, data_feature)
}
\arguments{
\item{onco_gene}{A charachter indicating a single onco_gene ID.}

\item{icp_gene}{A charachter indicating a single immune checkpoint ID.}

\item{data_expression}{A numeric matrix or data frame containing gene/protein expressions}

\item{data_feature}{A numeric matrix or data frame containing a single immune feature.}
}
\value{
a stratified boxplot
}
\description{
Generate a stratified boxplot for immune feature values
}
\details{
Feature data is stratified base on expression quartiles of onco_gene and icp_gene. High/Low categories include samples with expression values in lower/upper quartiles correspondingly. Samples with expression values in middle quartiles are discarded. For details of quartile calculation see get_quantile_rank function.

Pvalues are calculated using Wilcoxon test.

data_expression is formatted with genes/proteins as rows and samples/patients as columns.
For data_expression sample formats see \code{\link[imogimap]{sample_mRNA_data}}.

data_feature is formatted with samples/patients as rows and immune feature as single column.
For data_feature sample format see \code{\link[imogimap]{sample_Leukocyte_fraction_data}}.
}
\examples{
im_boxplot(onco_gene = "TGFB1",icp_gene="TNFSF4",
                  data_expression =  sample_mRNA_data,
                  data_feature = sample_Leukocyte_fraction_data)
}
\keyword{boxplot,}
\keyword{checkpoints}
\keyword{features,}
\keyword{immune}
