% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_boxplot_tcga_plot.R
\name{im_boxplot_tcga_plot}
\alias{im_boxplot_tcga_plot}
\title{Plot data from im_bloxplot_tcga}
\usage{
im_boxplot_tcga_plot(obj)
}
\arguments{
\item{obj}{a list obj produced by im_bloxplot_tcga}
}
\value{
a ggplot object
}
\description{
Plot data from im_bloxplot_tcga
}
\examples{
# FIXME


}
