% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cor.R
\name{im_cor}
\alias{im_cor}
\title{A correlation function}
\usage{
im_cor(gene_list, df_mrna, df_ict, df_lf)
}
\arguments{
\item{gene_list}{a list of charachters}

\item{df_mrna}{a formated mRNA data frame}

\item{df_ict}{a formated immune cell type fractions data frame}

\item{df_lf}{a formated Leukocyte fraction data frame}
}
\value{
a list of multiple dataframes of correlation coefficients and p.values
}
\description{
Calculates correlation with immune checkpoints
}
\examples{
im_cor(gene_list =  c("MCL1","TP53"),
                 df_mrna =  sample_mRNA_data,
                 df_lf = sample_Leukocyte_fraction_data,
                 df_ict = sample_immune_cell_fraction_data)
im_cor()
}
\keyword{features}
\keyword{immune}
