% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_netplot.R
\name{im_netplot}
\alias{im_netplot}
\title{Convert SIF to to igraph network}
\usage{
im_netplot(df, checkpoint, cohort, Immune_Feature, cutoff, seed)
}
\arguments{
\item{df}{A dataframe as outputed by im_syng_tcga or im_syng}

\item{checkpoint}{An optional character vector of immune checkpoints for color coding vertex. default is icp_gene_list}

\item{cohort}{An optional charachter string indicating a single TCGA disease.}

\item{Immune_Feature}{a charachter string indicating name of animmune feature.}

\item{cutoff}{A numeric indicating cut-off for absolute values of synergy scores.}

\item{seed}{A single value, interpreted as an integer, or NULL (see set.seed for details).}
}
\value{
A plotted igraph object showing network of synergistic (red) and antagonistic (blue) associations.
}
\description{
Convert SIF to to igraph network
}
\details{
im_netplot egenrate and plots network from synergy-score data frame. vertexes are gene names in dataframe and edges are synergistic (red) and antagonistic (blue) associations. Edges with absolute values lower that the cut-off wil be removed from the graph.

seed is an integer that is the starting point from which random numbers are generated in igraph. Seed value is used for the reproducibility of plot layout. different seed numbers will generate different
}
\examples{
df <- im_syng_tcga(cotarget = icp_gene_list,checkpoint = icp_gene_list,cohort = "brca")
im_netplot(df, Immune_Feature = "EMTscore",cutoff = 5, seed = 1234)
}
