% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_plot.R
\name{im_plot}
\alias{im_plot}
\title{Generate a stratified boxplot for immune feature values}
\usage{
im_plot(cotarget, checkpoint, data_expression, data_feature)
}
\arguments{
\item{cotarget}{A charachter indicating a single cotarget ID.}

\item{checkpoint}{A charachter indicating a single checkpoint ID.}

\item{data_expression}{A numeric matrix or data frame containing gene/protein expressions}

\item{data_feature}{A numeric matrix or data frame containing a single immune feature.}
}
\value{
a stratified boxplot
}
\description{
Generate a stratified boxplot for immune feature values
}
\details{
Feature data is stratified base on expression quartiles of cotarget and checkpoint. High/Low categories include samples with expression values in lower/upper quartiles correspondingly. Samples with expression values in middle quartiles are discarded. For details of quartile calculation see get_quantile_rank function.

data_expression is formatted with genes/proteins as rows and samples/patients as columns.
For data_expression sample formats see sample_mRNA_data.

data_feature is formated with samples/patients as rows and immune feature as single column.
For data_feature sample format see sample_Leukocyte_fraction_data.
}
\examples{
im_plot(cotarget = "TGFB1",checkpoint="TNFSF4",
                  data_expression =  sample_mRNA_data,
                  data_feature = sample_Leukocyte_fraction_data)
}
\keyword{boxplot,}
\keyword{checkpoints}
\keyword{features,}
\keyword{immune}
