% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_syng_tcga.R
\name{im_syng_tcga}
\alias{im_syng_tcga}
\title{A correlation function}
\usage{
im_syng_tcga(gene_list, cohort)
}
\arguments{
\item{gene_list}{a list of genes Hugo symbol}

\item{cohort}{a list of TCGA diseases}
}
\value{
a dataframe of correlation coefficient and p.values
}
\description{
This function allows you to calculate correlation with immune checkpoints
}
\examples{
im_syng_tcga(gene_list=c("TP53","TGFB1"),cohort=c("acc","gbm"))
im_syng_tcga()
}
\keyword{checkpoints}
\keyword{immune}
