% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_syng_tcga_shiny.R
\name{im_syng_tcga_shiny}
\alias{im_syng_tcga_shiny}
\title{Equivalent of im_syng_TCGA ro run with shiny app}
\usage{
im_syng_tcga_shiny(
  onco_gene,
  icp_gene,
  cohort,
  select_iap,
  method,
  ndatamin = 8,
  specificity,
  N_iteration_specificity = 1000,
  sensitivity,
  N_iteration_sensitivity = 100,
  sample_list,
  add_receptor_ligand = FALSE
)
}
\arguments{
\item{onco_gene}{A character vector of gene Hugo symbols.}

\item{icp_gene}{An optional character vector of immune checkpoint gene Hugo symbols.}

\item{cohort}{A list of TCGA diseases}

\item{select_iap}{An optional  character vector or numeric matrix or data.frame.}

\item{method}{A character string indicating which synergy score to be used. one of "max" or "independence".
Default is "max".}

\item{ndatamin}{minimum number of samples. Synergy score calculation will be skipped for matrices with number of rows less than ndatamin}

\item{specificity}{An optional logical indicating if specificity analysis should be done. Default is FALSE.}

\item{N_iteration_specificity}{Number of iterations for random sampling for specificity p.value calculation.
Default is 1000.}

\item{sensitivity}{An optional logical indicating if a sensitivity analysis should be done. Default is FALSE.}

\item{N_iteration_sensitivity}{Number of iterations for random sampling for sensitivity analysis.
Default is 100.}

\item{sample_list}{An optional character vector of TCGA samples barcodes indicating a subset of samples within a cohort.}

\item{add_receptor_ligand}{An optional logical indicating whether receptor_ligands pair should be added. Default is FALSE.}
}
\value{
A data.frame of synergy scores and bootstrapping p.values.
}
\description{
Takes a list of Tumor-intrinsic pathway(TIP) genes and returns their combinatorial association with immune checkpoint(ICP) genes by evaluating their synergistic impact on immune-associated phenotypes(IAP) using RNASeq2GeneNorm expressions as provided by \pkg{curatedTCGAData}.
}
\details{
If no icp_gene is specified, the default icp_gene_list will be used.
If add_receptor_ligand is TRUE, CellphoneDB database is used to add genes with known receptor-ligand interactions with either of icp_genes or onco_genes.
If select_iap is a character vector, it must be any sub-list of IAP names as listed in TCGA_immune_features_list. If a numeric matrix or data.frame, each column represents a user-defined IAP and must have a range between \code{[0,1]}. If select_iap is missing all IAPs listed in TCGA_immune_features_list will be considered for analysis.

For synergy score calculations all features are normalized to be on \code{[0,1]} range. For details of synergy score and significance pvalue calculations see \code{find_a_synergy} function.

A specificity p.value is computed using random sampling with replacement from two null models, generated from one of the two genes against a set of genes randomly selected from the genome. Two P-values are calculated for the synergistic interaction of the pair against the two null models. The highest of the two P-values is used to assess the specificity of the interaction against the whole genome. The number of randomly selected genes in each null model is determined by N_iteration_specificity.

Sensitivity (Robustness) score defined as normalized root mean square deviation of scores calculated over 70\% of samples, selected via random sampling. The number of sub-sample iterations is determined by N_iteration_sensitivity.

All barcodes in sample_list must be 15 character long and belong to the same cohort. When sample_list is provided, cohort should be the disease cohort that they belong to, otherwise only the first element of the cohort list will be used.
}
\examples{
im_syng_tcga_shiny(onco_gene=c("TGFB1","SERPINB9"), cohort=c("ucec"))

}
\keyword{Synergy}
\keyword{TCGA}
\keyword{bootstrapping,}
\keyword{checkpoint,}
\keyword{feature,}
\keyword{immune}
\keyword{scoring,}
