% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Round the numeric columns in a data.frame}
\usage{
round_df(df, digits = 5)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{the number of digits to round to}
}
\value{
a data.frame
}
\description{
Round the numeric columns in a data.frame
}
