% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_immune_cell_fraction_data}
\alias{sample_immune_cell_fraction_data}
\title{Sample data set for immune cell fraction data}
\format{
A data frame with of with 78 rows and 22 columns. Row names are a patient IDs and column names are cell types.
}
\source{
\url{https://cibersort.stanford.edu}
}
\usage{
sample_immune_cell_fraction_data
}
\description{
Data frame containing immune cell fractions for a subset of TCGA cohorts to serve as formatting guide for user-curated data or in test runs.
}
\keyword{datasets}
