% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_Leukocyte_fraction_data}
\alias{sample_Leukocyte_fraction_data}
\title{Sample Leukocyte fraction data}
\format{A data frame with of with 76 rows and 1 column. Row names are a patient IDs.}
\source{
\url{http://somewhere.important.com/}
}
\usage{
sample_Leukocyte_fraction_data
}
\description{
Data frame containing Leukocyte fractions for a subset of TCGA cohorts to serve as formatting guide for user-curated data or in test runs.
}
\keyword{datasets}
