% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_mRNA_data}
\alias{sample_mRNA_data}
\title{Sample mRNA expression data}
\format{A data frame with of with 19958 rows and 78 columns. Row names are gene Hugo IDs and column names are patient IDs.}
\source{
\code{\link[curatedTCGAData]{curatedTCGAData}}
}
\usage{
sample_mRNA_data
}
\description{
Data frame containing RNASeq2GeneNorm expressions for a subset of TCGA cohorts to serve as formatting guide for user-curated data or in test runs.
}
\keyword{datasets}
