% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_disease_list}
\alias{TCGA_disease_list}
\title{TCGA disease names}
\format{
A character vector of length 32
}
\source{
\code{\link[curatedTCGAData]{curatedTCGAData}}
}
\usage{
TCGA_disease_list
}
\description{
TCGA abbreviated disease names to be used as input in TCGA functions.
}
\keyword{datasets}
