% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_EMT}
\alias{TCGA_EMT}
\title{TCGA EMT scores}
\format{
A data frame with 9120 rows and 3 variables:
\describe{
\item{\code{Tumor_Sample_ID}}{character TCGA 15-charachter Tumor sample ID}
\item{\code{EMTscore}}{double  EMT score}
}
}
\source{
from Tong Pan
}
\usage{
TCGA_EMT
}
\description{
Epithelial–mesenchymal transition score for TCGA data. Higher EMT scores correspond to higher expression of mesenchymal genes.
}
\keyword{datasets}
