% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_immune_features_list}
\alias{TCGA_immune_features_list}
\title{TCGA immune features list}
\format{
A character vector of length 28
}
\usage{
TCGA_immune_features_list
}
\description{
Lists immune feature names that are default to TCGA functions.
}
\keyword{datasets}
