% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angio_score.R
\name{get_angio_score}
\alias{get_angio_score}
\title{Calculates Angiogenesis score
Calculates Angiogenesis score from a gene list}
\usage{
get_angio_score(pdf)
}
\arguments{
\item{pdf}{a formated mRNA data frame}
}
\value{
a dataframe of correlation coefficient and p.values
}
\description{
Calculates Angiogenesis score
Calculates Angiogenesis score from a gene list
}
\examples{
get_angio_score(pdf=sample_mRNA_data)
}
\keyword{checkpoints}
\keyword{immune}
