% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_syng_score.R
\name{Get_syng_score}
\alias{Get_syng_score}
\title{calculates a synergy score}
\usage{
Get_syng_score(fdata)
}
\arguments{
\item{fdata}{a formatted dataframe}
}
\value{
a dataframe with gene names and synergy scores
}
\description{
Generates coopretivity boxplots immune checkpoints
}
\examples{
im_syng(fdata=myformatteddata)
Get_syng_score()
}
\keyword{synergy}
