% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_syng_score.R
\name{get_syng_score}
\alias{get_syng_score}
\title{Finds a statistical synergistic score for a stratified data frame}
\usage{
get_syng_score(fdata, method)
}
\arguments{
\item{fdata}{A numeric data frame with 3 columns: A numeric indicating value of an immune feature and two integers, each interpreted as a coded label for the expression level of a gene. Each row of fdata is a different sample or experiment.}

\item{method}{a character string indicating which synergy score method to be used. one of "max" or "independence". Default is "max".}
}
\value{
A data frame containing synergy scores
}
\description{
Finds a statistical synergistic score for a stratified data frame
}
\details{
fdata is a numeric data frame with 3 columns: The first column contains the numeric value of a single immune feature. The second and third columns are any of the coded labels 1 or 4, representing low or high expression levels of two genes as outputted by get_quantile_rank. For details of quantile stratification see get_quantile_rank.
For synergy score calculations, fdata is stratified based on coded labels of the last two columns. Groups are labeled as low-low, low-high, high-low and high-high representing the expression levels of the corresponding two genes in each group. Median values of the feature column are then determined for each group.

Synergy scores are defined as the deviation of immune feature's median value in high-high group from the its expected median as estimated using one of the "max" or "independence" models. Lets define ma and mb as the difference between the levels of immune feature in low-low group and low-high/high-low groups respectively. The expected median of high-high group is then defined as max(ma , mb) in "max" model and as ma+mb-ma*mb in "independence" model. Synergy score is then calculated as the difference between mc and the expected median. Scores are called synergistic if mc exceeds the expected median. Antagonistic scores are ignored. The value of the synergy score are multiplied by 100 and a sign factor depending on the direction of the effect. By default synergy score will be zero if median values change in opposite directions. Differences in median values that are smaller that combined standard errors are considered as zero.
}
\examples{
dft <- TCGA_EMT$EMTscore
mydata<- cbind(feature=dft,gene1=sample(c(1,4),length(dft),replace = T),gene2=sample(c(1,4),length(dft),replace = T))
get_syng_score(mydata)

}
\seealso{
\code{\link[=get_quantile_rank]{get_quantile_rank()}}
}
\keyword{scoring}
\keyword{synergy}
