% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icp_gene_list}
\alias{icp_gene_list}
\title{Default list of immune checkpoints}
\format{A vector of characters
\describe{
Each character is a gene hugo symbol
}}
\source{
\url{https://www.cancerresearch.org/archive/clinical-accelerator-(2)/drug-agent-target-list}
}
\usage{
icp_gene_list
}
\description{
List of 29 immune checkpoints
}
\keyword{datasets}
