% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cor.R
\name{im_cor}
\alias{im_cor}
\title{Calculate Spearman correlation with immune checkpoints}
\usage{
im_cor(onco_gene, icp_gene, data_expression, data_feature, add_features)
}
\arguments{
\item{onco_gene}{A character vector of gene/protein IDs.}

\item{icp_gene}{An optional character vector of immune checkpoint gene/protein IDs.}

\item{data_expression}{A numeric matrix or data frame containing gene/protein expressions.}

\item{data_feature}{An optional numeric matrix or data frame containing immune features.}

\item{add_features}{An optional logical indicating if EMT score, angiogenesis score and IFNG expression should be added to immune features. Default is TRUE.}
}
\value{
A list of dataframes containing Spearman correlations and non-FDR adjusted probability values.
}
\description{
Calculate Spearman correlation with immune checkpoints
}
\details{
By default (if no icp_gene is specified), icp_gene_list will be used.

data_expression is formatted with genes/proteins as rows and samples/patients as columns.
For data_expression sample formats see see \code{\link[imogimap]{sample_mRNA_data}}.

data_feature is formatted with samples/patients as rows and immune features as columns.
For data_feature sample format see \code{\link[imogimap]{sample_Leukocyte_fraction_data}}.
}
\examples{
im_cor(onco_gene =  c("BRAF","CTLA4"),
                 icp_gene= c("CD274","CTLA4"),
                 data_expression =  sample_mRNA_data,
                 data_feature = sample_immune_cell_fraction_data,
                 add_features=TRUE)

}
\keyword{Spearman}
\keyword{checkpoints}
\keyword{correlation,immune}
\keyword{features,}
\keyword{immune}
