% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cor_tcga.R
\name{im_cor_tcga}
\alias{im_cor_tcga}
\title{Find Spearman correlation of co-target genes with immuno-oncology features using cbioportal data}
\usage{
im_cor_tcga(onco_gene, icp_gene, cohort)
}
\arguments{
\item{onco_gene}{A charachter vector of gene Hugo symbols.}

\item{icp_gene}{An optional charachter vector of immune checkpoint gene/protein IDs.}

\item{cohort}{a charachter vector of TCGA diseases}
}
\value{
a list of dataframes containing Spearman correlations and non-FDR adjusted probability values.
}
\description{
Find Spearman correlation of co-target genes with immuno-oncology features using cbioportal data
}
\details{
im_cor_tcga uses gene expressions from cbioportal data, 2018 tcga pancan atlas to find correlation and probaility values between onco_genes and immune checkpoints and immuno-oncology features as listed in TCGA_immune_features_list.

By default (if no icp_gene is specified), icp_gene_list will be used.

For TCGA disease list see TCGA_disease_list

A non-FDR-adjusted p.value is reported for each correlation value to allow for easier adjustments by user.
}
\examples{
im_cor_tcga(onco_gene=c("BRAF"),icp_gene=c("CD274","CTLA4"),cohort=c("acc","gbm"))
}
\keyword{correlation,}
\keyword{expression}
\keyword{features,}
\keyword{gene}
\keyword{immunouoncology}
