% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_cor_tcga.R
\name{im_cor_tcga}
\alias{im_cor_tcga}
\title{Finds Spearman correlation between an oncogene, an immune checkpoint and immune associated phenotypes.}
\usage{
im_cor_tcga(onco_gene, icp_gene, cohort, sample_list)
}
\arguments{
\item{onco_gene}{A character vector of gene Hugo symbols.}

\item{icp_gene}{An optional character vector of immune checkpoint gene/protein IDs.}

\item{cohort}{a character vector of TCGA diseases}

\item{sample_list}{An optional character vector of TCGA samples barcodes indicating a subset of samples within a cohort.}
}
\value{
a list of dataframes containing Spearman correlations and non-FDR adjusted probability values.
}
\description{
Finds Spearman correlation between an oncogene, an immune checkpoint and immune associated phenotypes.
}
\details{
im_cor_tcga uses NASeq2GeneNorm expression data, as provided by \code{\pkg{curatedTCGAData}}, to find correlation  between onco_genes and immune checkpoints and immuno-oncology features as listed in TCGA_immune_features_list.

By default (if no icp_gene is specified), icp_gene_list will be used.

For TCGA disease list see TCGA_disease_list

All barcodes in sample_list must be 15 character long and belong to the same cohort. When sample_list is provided, cohort should be the disease cohort that they belong to, otherwise only the first element of the cohort list will be used.

A non-FDR-adjusted p.value is reported for each correlation value to allow for easier adjustments by user.

All barcodes in sample_list must be 15 character long and belong to the same cohort. When sample_list is provided, cohort should be the disease cohort that they belong to, otherwise only the first element of the cohort list will be used.
}
\examples{
im_cor_tcga(onco_gene=c("BRAF"),icp_gene=c("CD274","CTLA4"),cohort=c("acc","gbm"))
}
\keyword{correlation,}
\keyword{expression}
\keyword{features,}
\keyword{gene}
\keyword{immuno-oncology}
