% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_netplot.R
\name{im_netplot}
\alias{im_netplot}
\title{Creates and plot an igraph network from dataframe}
\usage{
im_netplot(df, icp_gene, cohort, Immune_phenotype, cutoff, seed)
}
\arguments{
\item{df}{A dataframe as outputted by \code{\link{im_syng_tcga}} or \code{\link[imogimap]{im_syng}}}

\item{icp_gene}{An optional character vector of immune checkpoints for color coding vertex. default is \code{\link[imogimap]{icp_gene_list}}}

\item{cohort}{An optional character string indicating a single TCGA disease.}

\item{Immune_phenotype}{a character string indicating name of an immune feature.}

\item{cutoff}{A numeric indicating quantile cut-off for absolute values of synergy scores.}

\item{seed}{A single value, interpreted as an integer, or NULL (see set.seed for details).}
}
\value{
A plotted igraph object showing network of synergistic (red) and antagonistic (blue) associations.
}
\description{
Creates and plot an igraph network from dataframe
}
\details{
im_netplot uses \pkg{igraph} to construct and plot network from synergy score data frame. Immune-checkpoints/onco-genes are depicted as black/white vertices, and positive/negative synergistic interactions are depicted as red/blue edges. Thickness of an edge is determined by the absolute value of the score, and the size of each vertex is determined by its degree. Edges with absolute values lower that the cut-off will be removed from the graph.

cutoff=0 does not remove any edges. cutoff=1 removes all edges. Default is 0.95, which keeps scores with their absolute value  higher than 0.95 quantile.
Colored groups are constructed based on edge-between-ness membership. For more information see \code{\link[igraph]{edge.betweenness.community}}
seed is an integer that is the starting point from which random numbers are generated in igraph. Seed value is used for the reproducibility of plot layout. different seed numbers will generate different
}
\examples{
df <- im_syng_tcga(onco_gene = icp_gene_list,icp_gene = icp_gene_list,cohort = "acc")
im_netplot(df, Immune_phenotype = "EMTscore", seed = 1234)
}
\seealso{
\link{igraph}
}
