% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_netplot.R
\name{im_netplot}
\alias{im_netplot}
\title{Creates and plot an igraph network from dataframe}
\usage{
im_netplot(df, icp_gene, cohort, Immune_Feature, cutoff, seed)
}
\arguments{
\item{df}{A dataframe as outputed by im_syng_tcga or im_syng}

\item{icp_gene}{An optional character vector of immune checkpoints for color coding vertex. default is icp_gene_list}

\item{cohort}{An optional charachter string indicating a single TCGA disease.}

\item{Immune_Feature}{a charachter string indicating name of animmune feature.}

\item{cutoff}{A numeric indicating cut-off for absolute values of synergy scores.}

\item{seed}{A single value, interpreted as an integer, or NULL (see set.seed for details).}
}
\value{
A plotted igraph object showing network of synergistic (red) and antagonistic (blue) associations.
}
\description{
Creates and plot an igraph network from dataframe
}
\details{
im_netplot constructs and plots network from synergy score data frame. Immune-checkpoints/onco-genes are depicted as black/white vertices, and positive/negative synergistic interactions are depicted as red/blue edges. Thickness of an edge is determined by the absolute value of the score, and the size of each vertex is determined by its degree. Edges with absolute values lower that the cut-off wil be removed from the graph.

seed is an integer that is the starting point from which random numbers are generated in igraph. Seed value is used for the reproducibility of plot layout. different seed numbers will generate different
}
\examples{
df <- im_syng_tcga(onco_gene = icp_gene_list,icp_gene = icp_gene_list,cohort = "acc")
im_netplot(df, Immune_Feature = "EMTscore",cutoff = 5, seed = 1234)
}
