% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_netplot.R
\name{im_netplot}
\alias{im_netplot}
\title{Convert SIF to to igraph network}
\usage{
im_netplot(df, cohort, Immune_Feature, cutoff)
}
\arguments{
\item{sif}{a binary SIF as a data.frame with three columns:
"gene_A", "INTERACTION_TYPE", "gene_B"}

\item{directed}{a boolean weather the returned graph should be directed (DEFAULT: TRUE)}
}
\value{
a directed igraph network with interaction types
}
\description{
Convert SIF to to igraph network
}
\details{
Users are likely to run into issues if the input SIF has factor levels
}
