% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_plot_tcga.R
\name{im_plot_tcga}
\alias{im_plot_tcga}
\title{A correlation function}
\usage{
im_plot_tcga(gene1, gene2, cohort, Immune_Feauture)
}
\arguments{
\item{gene1}{a single gene name}

\item{gene2}{a single gene name}

\item{cohort}{a single TCGA disease}

\item{Immune_Feautue}{an immune feature name as listed in im_syng_tcga output.}
}
\value{
a list of multiple dataframes of correlation coefficients and p.values
}
\description{
Generates coopretivity boxplots immune checkpoints
}
\examples{
im_plot_tcga(gene1 = "TGFBR1",gene2="CSF1",
                  cohort="lihc", Immune_Feauture="Mast.Cells.Activated")
im_plot_tcga()
}
\keyword{boxplots}
