% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_syng.R
\name{im_syng}
\alias{im_syng}
\title{A correlation function}
\usage{
im_syng(gene_list, df_mrna, df_ict, df_lf)
}
\arguments{
\item{gene_list}{a list of charachters}

\item{df_mrna}{a formated mRNA data frame}

\item{df_ict}{a formated immune cell type fractions data frame}

\item{df_lf}{a formated Leukocyte fraction data frame}
}
\value{
a dataframes of bliss scores
}
\description{
Generates coopretivity boxplots immune checkpoints
}
\examples{
im_syng(gene_list =  c("TGBFR1","TGBFR2","ACVRL1","ENG"),
                  df_mrna =  sample_mRNA_data,
                  df_lf = sample_Leukocyte_fraction_data,
                  df_ict = sample_immune_cell_fraction_data)
im_syng()
}
\keyword{features}
\keyword{immune}
