% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_syng_tcga.R
\name{im_syng_tcga}
\alias{im_syng_tcga}
\title{Finds combinatorial association of immunotherapy co-targets with tumor intrinsic features as listed in TCGA_immune_features_list. Columns of the outputted data.frame contain synergy scores as well as p.values and a variances to measure strength, significance and sensitivity of the effects. Each row represents the combinatorial effect of two genes on an immune feature in a selected cohort.}
\usage{
im_syng_tcga(
  onco_gene,
  icp_gene,
  cohort,
  sample_list,
  method,
  feature,
  add_pvalue,
  N_iteration,
  sensitivity
)
}
\arguments{
\item{onco_gene}{A character vector of gene Hugo symbols.}

\item{icp_gene}{An optional character vector of immune checkpoint gene Hugo symbols.}

\item{cohort}{A list of TCGA diseases}

\item{sample_list}{An optional character vector of TCGA samples barcodes indicating a subset of samples within a cohort.}

\item{method}{A character string indicating which synergy score to be used. one of "max" or "independence". Default is "max".}

\item{N_iteration}{Number of iterations for random bootstrapping for p.value calculation and sensitivity analysis. Default is 1000.}

\item{sensitivity}{An optional logical indicating if a sensitivity analysis should be done. Default is FALSE.}

\item{data_feature}{An optional numeric matrix or data frame containing normalized immune features.}

\item{pvalue}{An optional logical indicating if a random bootstrapping value should be calculated. Default is FALSE.}
}
\value{
A data.frame of synergy scores and bootstrapping p.values.
}
\description{
Finds combinatorial association of immunotherapy co-targets with tumor intrinsic features as listed in TCGA_immune_features_list. Columns of the outputted data.frame contain synergy scores as well as p.values and a variances to measure strength, significance and sensitivity of the effects. Each row represents the combinatorial effect of two genes on an immune feature in a selected cohort.
}
\details{
im_syng_tcga uses RNASeq2GeneNorm expression data, as provided by \code{\pkg{curatedTCGAData}}, to find combinatorial association of immunotherapy co-targets and immune checkpoints with immuno-oncology features as listed in TCGA_immune_features_list

For synergy score calculations all features are normalized to be on \link{0,1} range. For details of synergy score calculations see get_syng_score function.

By default (if no icp_gene is specified), icp_gene_list will be used.

All barcodes in sample_list must be 15 character long and belong to the same cohort. When sample_list is provided, cohort should be the disease cohort that they belong to, otherwise only the first element of the cohort list will be used.

The optional data_feature must be normalized to have a range between \link{0,1}.

A p.value is computed using random bootstrapping with replacement from the distribution of synergy scores for each immune checkpoint and immune feature pair.

Sensitivity analysis reports a sensitivity index which is a variance in synergy score using 90\% of samples.
}
\examples{
im_syng_tcga(onco_gene=c("TP53","TGFB1"), cohort=c("acc","gbm"),add_pvalue=TRUE, N_iteration=1000)
}
\keyword{Synergy}
\keyword{TCGA}
\keyword{bootstrapping,}
\keyword{checkpoint,}
\keyword{feature,}
\keyword{immune}
\keyword{scoring,}
